[scheduler] Move the preset generation to a selector
